Inline PluginMeta into ChunkyPlugin.
